added util file
