+++ src/cnn_trainer/src/model/model.py
@@ -1,10 +1,7 @@
 import math
 import numpy as np
-import re
-import cv2
 import os
 
-from string import ascii_lowercase as LC
 from matplotlib import pyplot as plt
 from tensorflow.python.keras import layers
 from tensorflow.python.keras import models
@@ -12,80 +9,19 @@ from tensorflow.python.keras import optimizers
 from tensorflow.python.keras import backend
 from tensorflow.python.keras.utils import plot_model
 
+import util as util
+
 
 PATH = '/home/fizzer/ros_ws/src/cnn_trainer'
 PLATE_DIR = os.path.join(PATH, 'media', 'plates')
 MODEL_PATH = os.path.join(PATH, 'src', 'model')
-
-VALIDATION_SPLIT = 0.2
-
-
-# Get all files in a path
-def files_in_folder(path):
-    return [f for f in os.listdir(path) if os.path.isfile(os.path.join(path, f))]
-
-
-# Generate a one hot vector for a given character
-def one_hot(c):
-    vec = [0] * 36
-    try:
-        int(c)
-        index = int(c) + 26
-    except ValueError as e:
-        lowercase = c.lower()
-        index = LC.index(lowercase)
-    vec[index] = 1
-    return vec
-
-
-# Return an alphanumeric character from a given index
-def index_to_val(i):
-    abc123 = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789'
-    
-    return abc123[i]
-
-
-# Return the 4 partitions of a plate image and their associated one hot vectors
-def process_plate(my_file):
-    plate_path = os.path.join(PLATE_DIR, my_file)
-    plate_img = cv2.imread(plate_path)
-    
-    # crop into subsections (x dataset)
-    img1 = plate_img[49:349,45:150]
-    img2 = plate_img[49:349,145:250]
-    img3 = plate_img[49:349,347:452]
-    img4 = plate_img[49:349,447:552]
-    imgs = [img1,img2,img3,img4]
-
-    vecs = []
-    for i in range(4):
-        vecs.append(one_hot(my_file[6+i]))
-
-    return imgs, vecs
-
-
-# Return the datasets associated with the processed plates and their one-hot vectors
-def get_dataset():
-    plates = files_in_folder(PLATE_DIR)
-
-    X_images = []
-    Y_labels = []
-
-    for p in plates:
-        imgs, vecs = process_plate(p)
-        X_images.extend(imgs)
-        Y_labels.extend(vecs)
-
-    X_dataset = np.array(X_images) / 255  # normalize the data
-    Y_dataset = np.array(Y_labels)
-
-    return X_dataset, Y_dataset
 
 
 # Save a Keras model
 def save_model(model):
     model.save('keras/my_model')
     return
+
 
 # Load a Keras model
 def load_model():
@@ -113,20 +49,22 @@ def generate_model(lr):
 
 
 # Get a model either by generating a new one or load from local
-def get_model(X_dataset, Y_dataset, lr=1e-4, print_summary=False, new=False):  
+def get_model(X_dataset, Y_dataset, lr=1e-4, vs=0.2, print_summary=False, new=False):  
     
     print("Total examples: {}\nTraining examples: {}\nTest examples: {}".
             format(X_dataset.shape[0], 
-                   math.ceil(X_dataset.shape[0] * (1-VALIDATION_SPLIT)),
-                   math.floor(X_dataset.shape[0] * VALIDATION_SPLIT)))
+                   math.ceil(X_dataset.shape[0] * (1-vs)),
+                   math.floor(X_dataset.shape[0] * vs)))
     print("X shape: " + str(X_dataset.shape))
     print("Y shape: " + str(Y_dataset.shape))
 
     # Create a new model and save it or load one locally.
     if new:
+        print("compiling new model")
         conv_model = generate_model(lr)
         save_model(conv_model)
     else:
+        print("Loading local model")
         conv_model = load_model()
 
     conv_model.summary()
@@ -136,14 +74,8 @@ def get_model(X_dataset, Y_dataset, lr=1e-4, print_summary=False, new=False):
 
 # Predict a plate using a model
 def predict_plate(plate, model):
-    imgs, vecs = process_plate(plate)
+    imgs, vecs = util.process_plate(plate)
     dataset = np.array(imgs) / 255
-
-    char0 = np.array(imgs[0])
-    char1 = np.array(imgs[1])
-    char2 = np.array(imgs[2])
-    char3 = np.array(imgs[3])
-    total_plate = np.concatenate((char0,char1,char2,char3),axis=1)
 
     chars = []
 
@@ -156,22 +88,24 @@ def predict_plate(plate, model):
         y_predicted = model.predict(image)[0]
         index_predicted = np.argmax(y_predicted)
 
-        chars.append(index_to_val(index_predicted))
+        chars.append(util.index_to_val(index_predicted))
 
     print("Predicted:", chars) 
 
 
 def main():
     # PARAMETERS TO ADJUST
-    TRAIN = True
-    PRINT_HISTORY = True
+    TRAIN = False
+    PRINT_HISTORY = False
     LEARNING_RATE = 1e-4
+    VALIDATION_SPLIT = 0.2
     EPOCHS = 1
 
     # Get datasets and choose whether to generate a new model or load a model
-    X_dataset, Y_dataset = get_dataset()
-    conv_model = get_model(X_dataset, Y_dataset, lr=LEARNING_RATE, new=True)
+    X_dataset, Y_dataset = util.get_dataset()
+    conv_model = get_model(X_dataset, Y_dataset, lr=LEARNING_RATE, new=False)
     
+    # TODO: This can be its own function
     if TRAIN:
         history_conv = conv_model.fit(X_dataset, Y_dataset, 
                                     validation_split=VALIDATION_SPLIT, 
@@ -197,9 +131,9 @@ def main():
         plt.show()
 
     # Testing the model
-    plates = files_in_folder(PLATE_DIR)
-    print("Testing:", plates[0])
-    plate_to_test = plates[0]
+    plates = util.files_in_folder(PLATE_DIR)
+    print("Testing:", plates[300])
+    plate_to_test = plates[300]
 
     predict_plate(plate_to_test, conv_model)
